+++ src/modals/index.ts
@@ -4,3 +4,4 @@ export * from './profile';
 export * from './media';
 export * from './activity';
 export * from './milestone';
+export * from './update';